requirements.txt
aiohttp>=3.9.0
python-telegram-bot>=20.0
feedparser>=6.0.0
